refactor: Simplify EventCreationDialog by delegating to specialized components
- Refactor EventCreationDialog.kt to act as a facade, delegating to the specialized EventCreationDialog component for improved maintainability and organization.
- Remove legacy UI logic and streamline the structure for better clarity and consistency.
